+++ 1-main.c
@@ -10,7 +10,7 @@
  *
  * Return: Pointer to the first element of the created list. NULL on failure
  */
-listint_t *create_listint(const int *array, size_t size)
+listint_t *create_listint(const int *array, int size)
 {
 	listint_t *list;
 	listint_t *node;
@@ -42,7 +42,7 @@ int main(void)
 {
 	listint_t *list;
 	int array[] = {19, 48, 99, 71, 13, 52, 96, 73, 86, 7};
-	size_t n = sizeof(array) / sizeof(array[0]);
+	int n = sizeof(array) / sizeof(array[0]);
 
 	list = create_listint(array, n);
 	if (!list)
