+++ print_array.c
@@ -7,9 +7,9 @@
  * @array: The array to be printed
  * @size: Number of elements in @array
  */
-void print_array(const int *array, size_t size)
+void print_array(const int *array, int size)
 {
-	size_t i;
+	int i;
 
 	i = 0;
 	while (array && i < size)
